Switch score type to integer. No more use of long in this project.

+++ AsteroidMiner/src/ch/kanti_wohlen/asteroidminer/Player.java
@@ -10,7 +10,7 @@ public interface Player extends Disposable {
 
 	public int getPlayerNumber();
 
-	public long getScore();
+	public int getScore();
 
 	public void doInput();
 }
